+++ src/main/java/com/erick/nutricontrol/service/impl/ScheduleRuleServiceImpl.java
@@ -3,11 +3,14 @@ package com.erick.nutricontrol.service.impl;
 import com.erick.nutricontrol.dto.scheduleRule.ScheduleRuleDetailDTO;
 import com.erick.nutricontrol.dto.scheduleRule.ScheduleRuleRequestDTO;
 import com.erick.nutricontrol.dto.scheduleRule.ScheduleRuleUpdateDTO;
+import com.erick.nutricontrol.exception.BadRequestException;
 import com.erick.nutricontrol.exception.ConflictException;
 import com.erick.nutricontrol.exception.NotFoundException;
 import com.erick.nutricontrol.mapper.ScheduleRuleMapper;
 import com.erick.nutricontrol.model.ScheduleRule;
 import com.erick.nutricontrol.repository.ScheduleRuleRepository;
+import com.erick.nutricontrol.security.user.model.User;
+import com.erick.nutricontrol.security.user.repository.UserRepository;
 import com.erick.nutricontrol.service.ScheduleRuleService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -24,10 +27,15 @@ import java.util.List;
 public class ScheduleRuleServiceImpl implements ScheduleRuleService {
   private final ScheduleRuleRepository repository;
   private final ScheduleRuleMapper mapper;
+  private final UserRepository userRepository;
 
   @Override
   @Transactional
-  public ScheduleRuleDetailDTO addScheduleRule(ScheduleRuleRequestDTO dto) {
+  public ScheduleRuleDetailDTO addScheduleRule(String username, ScheduleRuleRequestDTO dto) {
+    if(dto.startTime().isAfter(dto.endTime()) || dto.startTime().equals(dto.endTime())) {
+      throw new BadRequestException("Invalid start and end time");
+    }
+    User user = userRepository.findByUsername(username).orElseThrow(() -> new NotFoundException("User not found"));
     List<ScheduleRule> list = repository.findByDayOfWeek(dto.dayOfWeek());
     LocalTime newStart = dto.startTime();
     LocalTime newEnd = dto.endTime();
@@ -39,29 +47,40 @@ public class ScheduleRuleServiceImpl implements ScheduleRuleService {
       }
     }
     ScheduleRule scheduleRule = mapper.toEntity(dto);
+    scheduleRule.setAdmin(user);
     scheduleRule = repository.save(scheduleRule);
     return mapper.toDetailDTO(scheduleRule);
   }
 
   @Override
-  public Page<ScheduleRuleDetailDTO> listScheduleRules(Pageable pageable) {
-    Page<ScheduleRule> list = repository.findAll(pageable);
-    if (list.isEmpty()) {
+  public Page<ScheduleRuleDetailDTO> listScheduleRulesByAdmin(String username, Pageable pageable) {
+    User user = userRepository.findByUsername(username).orElseThrow(()-> new NotFoundException("User not found"));
+    Page<ScheduleRule> page = repository.findByAdmin(user, pageable);
+    if(page.isEmpty()){
       return Page.empty();
     }
-    return list.map(mapper::toDetailDTO);
+    return page.map(mapper::toDetailDTO);
   }
 
   @Override
-  public ScheduleRuleDetailDTO getScheduleRuleById(Long id) {
+  public ScheduleRuleDetailDTO getScheduleRuleById(String username, Long id) {
+    User user = userRepository.findByUsername(username).orElseThrow(()-> new NotFoundException("User not found"));
     ScheduleRule scheduleRule = repository.findById(id).orElseThrow(() -> new NotFoundException("Schedule Rule Not Found"));
+    if(!user.equals(scheduleRule.getAdmin())) throw new BadRequestException("Invalid user");
     return mapper.toDetailDTO(scheduleRule);
   }
 
   @Override
   @Transactional
-  public ScheduleRuleDetailDTO updateScheduleRule(Long id, ScheduleRuleUpdateDTO dto){
+  public ScheduleRuleDetailDTO updateScheduleRule(String username, Long id, ScheduleRuleUpdateDTO dto){
+    if(dto.startTime().isAfter(dto.endTime()) || dto.startTime().equals(dto.endTime())) {
+      throw new BadRequestException("Invalid start and end time");
+    }
+    User user = userRepository.findByUsername(username).orElseThrow(()-> new NotFoundException("User not found"));
     ScheduleRule scheduleRule = repository.findById(id).orElseThrow(() -> new NotFoundException("Schedule Rule Not Found"));
+    if(!user.equals(scheduleRule.getAdmin())) {
+      throw new BadRequestException("Invalid user");
+    }
     mapper.updateEntityFromDto(dto, scheduleRule);
     scheduleRule = repository.save(scheduleRule);
     return mapper.toDetailDTO(scheduleRule);
@@ -69,8 +88,12 @@ public class ScheduleRuleServiceImpl implements ScheduleRuleService {
 
   @Override
   @Transactional
-  public void deleteScheduleRuleById(Long id) {
+  public void deleteScheduleRuleById(String username, Long id) {
+    User user = userRepository.findByUsername(username).orElseThrow(()-> new NotFoundException("User not found"));
     ScheduleRule scheduleRule = repository.findById(id).orElseThrow(() -> new NotFoundException("Schedule Rule Not Found"));
+    if(!user.equals(scheduleRule.getAdmin())) {
+      throw new BadRequestException("Invalid user");
+    }
     repository.delete(scheduleRule);
   }
 }
